+++ NEWS.md
@@ -1,4 +1,6 @@
 # fmriqa 0.2.0
+* Added an argument for the polynomial detrending order and changed the default
+from 2 to 3.
 * Improved k-space spike detection plot limits for very stable data.
 * COG calculations are now performed on a thresholded image to reduce bias from
 inhomogenity.